# Enable SkipLocalsInit for .NET6