[SkipLocalsInit] can be used to improve performance - https://github.com/dotnet/roslyn/blob/master/docs/features/skiplocalsinit.md

Runtime applies this attribute on all their assemblies for 5.0 - https://github.com/dotnet/runtime/search?q=SkipLocalsInit